+++ source/virtual/vgame.c
@@ -61,7 +61,7 @@
 #define NAME_NDS_DATADIR    "data"
 
 
-static u32 vgame_type = 0;
+static u64 vgame_type = 0;
 static u32 base_vdir = 0;
 
 static VirtualFile* templates_cia   = (VirtualFile*) VGAME_BUFFER; // first 56kb reserved (enough for 1024 entries)
@@ -634,8 +634,8 @@ bool BuildVGameFirmDir(void) {
     return true;
 }
 
-u32 InitVGameDrive(void) { // prerequisite: game file mounted as image
-    u32 type = GetMountState();
+u64 InitVGameDrive(void) { // prerequisite: game file mounted as image
+    u64 type = GetMountState();
     
     vgame_type = 0;
     offset_firm  = (u64) -1;
